several bugs fixed, message logcat changed from String to List<String>

+++ clap-server/clap-server/clap-server-db/src/main/java/com/noveogroup/clap/dao/impl/GenericDAOImpl.java
@@ -51,7 +51,9 @@ public abstract class GenericDAOImpl<T, ID extends Serializable> implements Gene
     @Override
     public void removeById(final ID id) {
         final T obj = entityManager.find(persistentClass, id);
-        entityManager.remove(obj);
+        if (obj != null) {
+            entityManager.remove(obj);
+        }
     }
